Revert initialisation of nIndex to 1 instead of 0
1 is annoying because people could get confused that an axis should
have Axis 1 as it's master. However it's a pain to try to work with 0
as the GearInMultiMaster function block needs to have all 4 axis refs filled.
An alternative is to pass an Empty Axis ref but I found it to be more trouble than
it's worth. So we can initialise with axis 1 and a gear ratio of 0 means it
won't actually have any affect. This is wear the status bit bCoupledGear1 etc.
is useful.