Logic changes to find a working solution.
The code is constantly checking the coupled status and putting
this into fSlaveAxisRatio in the struct. It only resets things to zero when
a gear out command is issued. This could be troublesome in the future.
The axes use this array fSlaveAxisRatio to check whether there are other limits that
it has to worry about apart from it's own.