Add "widget" in Visualization for GearInMultiMaster
The function block can have up to four masters so there are
four rows with a space to select the master axis and a gear ratio.
A gear ratio of 0 means the master won't have any affect. This needs to
be either fixed so one can select "No Master" or documented somewhere.