MBP-131: Update homing routines based on Beckhoff Example code.
Previously each homing routine had it's own state machine which
resulted in alot of duplicate code. Now all homing routines use
the same state machine.

The parts of the state machine which differ between homing routines
are selected via a case statement. The case statement calls a
the appropriate ACT to execute that specific detection FB of the
chosen homing routine. The state machine acts on bits which are
input/output to the FBs.

Previously FB_StepLimitSwitch, FB_StepAbsoluteSwitch and
FB_StepReferencePulse were used. FB_StepLimitSwitch resulted
in alot of issues when testing on hardware; axis would get stuck
in a state where it would oscilate back and fourth over the limit
switch. These three routines have been changed to the Detection
version. The difference is these Detection FBs don't do the set
position, they return a fRecordedPosition and the set position
is done separatly.

Use VelocityToCam when moving to cam and VelocityFromCam when looking
for the final sensor. Get these values from NC.

Feature added to input a fFinishDistance to fbFinishHoming. This
will move the axis a relative distance after homing has been
completed. fFinishDistance added to stAxisConfig.stConfig.

Add buttons on the VISU for all standard homing routines.
Prevously there was only buttons for two of the homing routines.
Now there are buttons for all the single speed routines. I've left off
the 2 speed ones for the moment because there was conjecture around
whether they were required.

Add logic to keep axis enable TRUE after limit switch hit until it has slowed.
Add bWaitForStoppingAfterLimitHit. There is a short timer, 100ms, so that once
a limit is hit there is a buffer for the axis to start slowing down. After
100ms the bit will stay high while the axis is decelerating. The timer
just bridges the gap between the limit switch hit and the
axis beginning to slow. This means axes need to be commissioned
properly in order for a limit switch hit to have low enough
speed and high enough deceleration to be able to stop before
hitting anything.

Logic change for stopping: If fbHome.bBusy is TRUE for the homing
block and a stop is issued then FB_Axis will do fbHalt instead
of fbStop. fbStop creates errors in the advanced homing blocks
where as fbHalt does not.

Modify logic that registers a warning: fbHome.bCommandAborted = TRUE
latches a warning in fbAxis even though it dissapoears in the
function block. This is then cleared on the next command execute.
Not sure if this is a good way to operate or whether it should
be a warning at all.

Remove duplicate code for Gearing Status - not sure how it got there.

Remove fbStopRiseEdge and fbResetRiseEdge -> unused.

Remove widget for bToggle, not required to current operation of FB_Axis.

Shuffle items around to be more space efficient. Move the widget for
single axis gearing to lower right as it should be removed after we
are confident with multi axis gearing.