Prepare the filter to ignore/remove LineIDs
LineIDs are put into the xml files by the IDE
to store information about lines in the source code for debugging
purposes.
These LineIDS often create unnecessary diffs in `git diff`
and/or merge conflicts.

TwinCAT can be configured to store them in a .dbg file:
This commits prepares the filter(s) to clean up the files, the cleanup
will be done in the next commit

Note: Credits to frm2, from where we copied filter/