Change the handling of SW and HW temperature values
Change the values of the MAX nad MIN_TEMP_LIMIT to scaled values.
Move the selection between HW temperature and SW temperature to the Init state
fTemperatureValue will be the local variable containing the tempearture value,
eitehr the HW or the simulated one.
Use just fTemperatureValue for all other calcualtions and assigments.
If the FB is disabled the display temperature will be 0.0

modified:   POUs/FB_AxisTemperature.TcPOU