Added delay timers for out of pressure errors
In the previous version of the code, the delay timers for the air pressure errors were not working properly.
The axus would go into error as soon as the air pressure would go out of range.
Now, the error activation is delayed for certain time, and this is allowing air pressure to return to allowed range without error activation.