Corrects the wrong links in the MainVisu. By opening the MainVisu.TcVIS, these are the links that were wrong: <v n="InputBoxVariable">"GVL.axes[MAIN.hmiAxisSelection].config.stMultiMasterAxis[1].fRatio"</v> <v n="InputBoxVariable">"GVL.axes[MAIN.hmiAxisSelection].config.stMultiMasterAxis[2].fRatio"</v> <v n="InputBoxVariable">"GVL.axes[MAIN.hmiAxisSelection].config.stMultiMasterAxis[3].fRatio"</v> <v n="InputBoxVariable">"GVL.axes[MAIN.hmiAxisSelection].config.stMultiMasterAxis[4].fRatio"</v> <v n="InputBoxVariable">"GVL.axes[MAIN.hmiAxisSelection].config.fAcceleration"</v> <v n="InputBoxVariable">"GVL.axes[MAIN.hmiAxisSelection].config.fDeceleration"</v> <v n="InputBoxVariable">"GVL.axes[MAIN.hmiAxisSelection].config.fPosition"</v>
It was find out that probably it was working before because in the configuration of the OnMouseClick, after clicking in Write a Variable, the item "Use text output variable" was checked.
It was fixed by taking the Write a Variable away, and putting it back, on the dialog box of the OnMouseClick input configuration.

Now in the MainVisu.TcVIS the previously wrong links are like this:
<v n="InputBoxVariable">""</v>