Add latest version of MAINVISU, missing the Coe Parameters part
When merging master to this branch the MAINVISU stopped working.
When tryin gto checkout th eMAINVISU from the pre-Merge branch it was
also not working (maybe library versions conflict)
We currently hav ethe latest version of the MAINVISU but the the part
added for the read/write COE parameters is missing and will have
to be added manually again before merging to master.

modified:   VISUs/MainVisu.TcVIS