Print log before timeout of new print
Normally a timeout is set before a new log message of the same kind is
allowed to be printed. If an error would occur with a slave, a log
message would be printed with that error and information about the slave
number. If, during the timeout, an additional slave has a problem, a new
message was previously not printed until after the timeout. Allow
printing to log if extra message changes in between each loop.