Use local scope
Moves the declaration of variables/function blocks, which are only
accessed in a certain method, from the general declaration in the FB to
the declaration inside that method in order to limit its scope. Uses the
VAR_INST specifier to allow TwinCAT to remember the previous state of
the variable/fb.