MBP-118: Redefine enabled and motion prevention
Changes the definition of bLocalEnabled to reflect the status of
enabled and output of brake (if external brake control).

Uses the original stAxis.stStatus.bEnabled in the mControl_ExternalBrake
instead of the redefined bLocalEnabled.

Removes the check of stAxis.stStatus.bEnabled in mMove_MoveAbsolute,
which could prevent a motion command from executing if the brake was not
released. Calls instead the fbPower.Enable_Negative/Positive to prevent
motion if brake not released.