Rename FB to match MCU5001a. Update Hw links and varible names
Becuase this FBis now specific for the MCU5001a, the FB was renamed.
The AbnormalLED does not exist anymore, remove it from the alarm array
The terminal names have chaged to match the ePlan drawings therefore,
the hardware links have to be changed to match the new names.
Comment the variables htat are not used for this cabinet and delete
the variables hat not exist anymore.