MBP-118: Implement mControl_CommandHandler
Implements a CommandHandler to facilitate handling of incoming commands.
This introduces a layer inbetween the commands and the motion functions.
With e.g. an airpad and external brake involved, this design pattern
will help with holding the set command until ready to execute (e.g.
airpad pressurized or brake released). Fur future possibilities of
automatic enabling/disabling of axis, this will also be helpful.