MBP-118: Add and move variables into stSTatus
Add a new variable to the stStatus to highlight if the CommandHandler is
buffering a command (i.e. waiting for a move permit from e.g. an airpad
or external brake).
Moves the bAxisReadyToMove into the stStatus structure and updates the
code accordingly.
Separates the acting on the timeout in the command handler from standard
stop/halt to facilitate setting the timeout alarm.