MBP-118: Add variables for manual ctrl of airpad
Extends the stControl with two variables two allow for manual control of
the airpad. One variable is used to enable manual control, the other to
control the pressurization when manual control i selected.

Takes a snapshot of the bMoveRequest variable when selecting manual
control, to continue the current status of the airpad when switching to
manual control.