Create abstract monitoring base class
Creates an abstract base class which all monitoring functions will
inherit from.

The base class includes an implementation of all common
methods, as well as two declared abstract methods, mLink and
mSetResetAlarmsLeds. The classes who inherit from the base class will
need to implement the abstract methods.

To facilitate the OOP solution, several internal help variables were
created in the base class which can be accessed through getter/setter
properties. The inheriting classes declare and link all I/O variables.
A number of these I/O variables will need to be linked to the
corresponding property of the base class in order for the logic to work
as expected.