Refactor Acknowledge
Moves the handling of the Acknowledge of an alarm from being executed in
the end of the mSetResetAlarmsLeds method to be handled directly in
mSetReset, i.e. the same function as where the auto reset is being
performed. The refactoring means that the inheriting FBs don't need to
call the Acknowledge handling explicitly any longer.